requirements.txt
accelerate==1.7.0
bitsandbytes==0.46.0
google-genai==1.17.0
pillow==11.2.1
python-dotenv==1.1.0
requests==2.32.3
streamlit==1.45.1
torch==2.7.0
transformers==4.52.3
